// src/app/core/api/generated/schemas/carBrandModelControllerListSortBy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */

export type CarBrandModelControllerListSortBy =
  (typeof CarBrandModelControllerListSortBy)[keyof typeof CarBrandModelControllerListSortBy];

export const CarBrandModelControllerListSortBy = {
  newest: "newest",
  oldest: "oldest",
  name: "name",
} as const;