// src/app/core/api/generated/schemas/updateCarBrandModelDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */

export interface UpdateCarBrandModelDto {
  car_brand_id?: number;
  model_name?: string;
  model_engine?: string;
  model_image?: string;
  is_active?: boolean;
  is_deleted?: boolean;
}